+++ src/engine/evaluation/imbalance_eval.rs
@@ -95,7 +95,22 @@ impl ImbalanceEvalTrait for Board {
 mod tests {
 
     use crate::engine::board::fen::FenTrait;
+    use crate::engine::board::structures::color::{BLACK, WHITE};
+    use crate::engine::evaluation::init_eval::InitEvalTrait;
     use crate::engine::evaluation::test_evaluation::SF_EVAL;
 
     use super::*;
+
+    // NOTE: 3. IMBALANCE [TEST: WORKS]
+    #[test]
+    fn imbalance_test() {
+        for obj in &SF_EVAL {
+            let mut board = Board::read_fen(obj.fen);
+            board.init();
+            board.imbalance(WHITE);
+            board.imbalance(BLACK);
+
+            assert_eq!(board.calculate_score(), obj.imbalance);
+        }
+    }
 }
