+++ src/engine/evaluation/tempo_eval.rs
@@ -20,7 +20,19 @@ impl TempoEvalTrait for Board {
 mod tests {
 
     use crate::engine::board::fen::FenTrait;
+    use crate::engine::evaluation::init_eval::InitEvalTrait;
     use crate::engine::evaluation::test_evaluation::SF_EVAL;
 
     use super::*;
+
+    #[test]
+    fn tempo_test() {
+        for obj in &SF_EVAL {
+            let mut board = Board::read_fen(obj.fen);
+            board.init();
+            board.tempo(board.color());
+
+            assert_eq!(board.calculate_score(), obj.tempo);
+        }
+    }
 }
